*** empty log message ***
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@434 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ modules/dx_hw/dx_window.c
@@ -300,7 +300,7 @@ typedef BOOL (WINAPI* typSetLayeredWindowAttributes)(HWND,COLORREF,BYTE,DWORD);
 static void SetWindowless(GF_VideoOutput *vout, HWND hWnd)
 {
 	const char *opt;
-	u8 a, r, g, b;
+	u32 a, r, g, b;
 	COLORREF ckey;
 	typSetLayeredWindowAttributes _SetLayeredWindowAttributes;
 	HMODULE hUser32;
